[bazel] add lowrisc_opentitan as a dependency
This adds the lowrisc_opentitan repository as a Bazel dependency in order
to build opentitanlib as a dependency for an FPGA DUT shim layer that
will be used for running provisioning flow tests in the lab.

Additionally, this adds skylib, rules_python, and rules_rust as external
dependencies, as these are also required for building opentitanlib.

Lastly, this updates the CRT and misc_linters dependencies, as this is
required to build opentitanlib as well.

Signed-off-by: Tim Trippel <ttrippel@google.com>