[spm] Implement EndorseCerts
The `EndorseCerts` RPC takes a list of certificate bundles, each
containing a signing key and a list of TBS certificates to sign. The SPM
hashes the TBS based on the signing key configuration, calculates the
signature, and generates a signed certificate in DER format.

This commit also introduces an unittest based on the opentitan
orchestrator generated output, as well as pre-configured DICE CA private
key and root CA.

This implementation only supports ECDSA keys.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>