[pa/spm] Add support for deriving new symmetric keys.
This change updates the `GenerateSymmetricKeys` API to add support for
symmetric key generation. Symmetric keys are ephemeral, thus the caller
is required to request a `wrapped_seed` with the result which will
return the seed encrypted by a public key.

The wrapped seed can be unwrapped in an offline HSM to retrieve the seed
and re-grenerate the original key.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>