Create syncer job between database and registry
This worker is used to sync records present in the database into the
provided registry. It will periodically retrieve all unsynced records
from the database, then send them to the registry in a batch request.
Only the successful records will be listed as synced, the other ones
will be retried in a later run.

This worker runs in the background as a goroutine and can be
configured to run with different frequency and records per run.