[spm] Remove SKU from diversification string in GenerateTokens
The GenerateTokens function is expected to be used in both `CP` and `FT`
stages, and in some cases, the derived tokens need to be reproducible.
For example, the Wafer Authentication Secret is injected into the Device
Under Test (DUT) in the `CP` stage, and then later reproduced by the HSM
in the `FT` stage.

This change removes the use of the SKU name as a prefix in the
diversification string because the customer SKU name may not be known by
`CP`. For example, a Silicon Creator may want to process a batch of
wafers in `CP` and later break it down into SKU allocations during
`FT`. Removing the SKU identifier allows the Silicon Creator to use a
reproducible derivation string that is not dependent on the SKU label.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>