Add JSON configuration for HTTPRegistry
This change introduces a new way of passing configuration to the HTTP
Registry client. A JSON file can be passed with all the relevant
configuration. I also included a wrapper constructor which takes a JSON
file path and it will handle the reading and parsing.

I also updated the flags in ProxyBuffer server, as the previous ones are
now irrelevant.