[hsm] Add support for Gem engine.
This commits enables the use of the gem engine with Thales HSMs. The
test infrastructure uses the `OTPROV_USE_GEM_ENGINE=true` enable it.

The token_init.sh script was moved under //config:token_init.sh. The
script is now called in prod mode, but restristed to only run certgen.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>