[proto] add num perso TLV objects field DeviceData
A perso blob is contained in the proto message that gets sent to the
device registry. The "perso blob" contains a TLV data structure of all
assets provisioned on the DUT, including certs and seeds. To make it
easier to parse this data structure, we include the number of objects
contained within it in the message that is sent to the registry.

Signed-off-by: Tim Trippel <ttrippel@google.com>