[ate] make seed data type generic
After https://github.com/lowRISC/opentitan/pull/27423, there are two
seed types that can be exported from the DUT during personalization:
1. "dev_seed", which is a seed that has a particular data format:
https://github.com/lowRISC/opentitan/blob/c4214fe7dfeb5ea3e203d14620bd4af87c451ba3/sw/device/silicon_creator/manuf/base/perso_tlv_data.h#L59
and,
2. a "generic" seed.

Both seed types can be represented in the TLV perso blob structure with
a simple header that contains the size and type (dev seed or generic),
followed by the data bytes.

This simplifies the ATE DLL to be able to represent both seed types with
the same `seed_t` object type.

Signed-off-by: Tim Trippel <ttrippel@google.com>