[hsm] use backwards compatible key wrap mechanism
There are two versions of the vendor-specific AES key wrapping
mechanism:
1. VendorThalesAesKw, and
2. VendorThalesAesKwp.

The former is only capatible with keys that are multiples of 64-bits,
while the latter is compatible across all key sizes ("p" stands for
padded version). However, only the former (VendorThalesAesKw) is
supported between specific version of the HSM we have deployed.
Therefore we reconfigure the prod deployment to use this key
wrapping/unwrapping mechanism.

Note, when using the VendorThalesAesKw mechanism, it is imparative the
CKA_VALUE_LEN attribute is set when importing a wrapped secret into the
HSM.

Signed-off-by: Tim Trippel <ttrippel@google.com>