[spm]: Add device data verification during registration
This change introduces a verification step for device data during the
device registration process in the Provisioning Appliance (PA).

The SPM now exposes a `VerifyDeviceData` RPC that the PA calls.
This RPC validates the device's certificate chain, which is passed in a
personalization TLV blob.

The verification logic in the SPM supports multiple DICE certificate
chain configurations, controlled by a new `CertChainDiceLeaf` SKU
attribute. This allows for flexible certificate hierarchies, such as:

- Root CA -> UDS ICA -> UDS (leaf)
- Root CA -> UDS ICA -> UDS -> CDI_0 (leaf)
- Root CA -> UDS ICA -> UDS -> CDI_0 -> CDI_1 (leaf)

The `dututils` in the ATE framework have been updated to emulate the
generation of these certificate chains on the DUT for testing purposes.
The load test now utilizes this functionality to test the end-to-end
flow.

Additionally, FPGA-specific identifiers have been added to the
device ID proto definitions.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>