[loadtest] Simulate manufacturing flow.
This change refactors the manufacturing flow to simulate the CP and FT
process. The test now uses the `dututils.Dut` object to maintain a
stateful representation of the device under test.

The `dututils.NewDut` method was updated to include the inintialization
of TBS certificates. This was needed to avoid race conditions when
accessing the HSM during loadtest operation.

The loadtest program also reports a calculated chip/hour rate as well as
more detailed error messages.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>