[ate] add `dut_rx_spi_frame_t` struct
Previously, we used one struct type for sending data to / from the DUT
over the SPI console. However, the size of a SPI console data frame that
is TXed by the DUT to the ATE is a maximum size of 2020 bytes, while the
size of a SPI console data frame that is RXed by the DUT from the ATE is
a maximum size of 256 bytes.

Therefore, we add a stuct to represent a SPI console frame that is RXed
by the DUT too to facilitate ATE integrations.

Signed-off-by: Tim Trippel <ttrippel@google.com>