[pa,spm,ate] add scaffolding to verify hash of all certs
After writing all certificates to flash, the DUT reads out each
certificate and computes a SHA256 hash over them. It then sends this
hash to the host so the provisioning-infrastructure can validate the
certificates were installed successfully into non-volatile storage on
the DUT.

To validate this hash, we update the RegisterDevice API to validate the
hash of all certs written to flash.

Signed-off-by: Tim Trippel <ttrippel@google.com>